+++ netreach-v4/iniparser/iniparser_wrapper.c
@@ -75,22 +75,22 @@ void IniparserWrapper::get_client_mac(uint8_t *macaddr) {
 
 // Server
 
-size_t IniparserWrapper::get_split_num() {
+uint32_t IniparserWrapper::get_split_num() {
 	int tmp = iniparser_getint(ini, "other:split_num", -1);
 	if (tmp == -1) {
 		printf("Invalid entry of [other:split_num]: %d\n", tmp);
 		exit(-1);
 	}
-	return size_t(tmp);
+	return uint32_t(tmp);
 }
 
-size_t IniparserWrapper::get_server_num() {
+uint32_t IniparserWrapper::get_server_num() {
 	int tmp = iniparser_getint(ini, "server:server_num", -1);
 	if (tmp == -1) {
 		printf("Invalid entry of [server:server_num]: %d\n", tmp);
 		exit(-1);
 	}
-	return size_t(tmp);
+	return uint32_t(tmp);
 }
 
 short IniparserWrapper::get_server_port() {
